+++ R/combine_plots.R
@@ -61,11 +61,14 @@
 #' @importFrom cowplot draw_label
 #' @importFrom grid unit
 #'
+#' @references
+#' \url{https://cran.r-project.org/web/packages/ggstatsplot/vignettes/combine_plots.html}
+#'
 #' @examples
-#' 
+#'
 #' # loading the necessary libraries
 #' library(ggplot2)
-#' 
+#'
 #' # preparing the first plot
 #' p1 <-
 #'   ggplot2::ggplot(
@@ -74,7 +77,7 @@
 #'   ) +
 #'   geom_point() +
 #'   labs(title = "setosa")
-#' 
+#'
 #' # preparing the second plot
 #' p2 <-
 #'   ggplot2::ggplot(
@@ -83,7 +86,7 @@
 #'   ) +
 #'   geom_point() +
 #'   labs(title = "versicolor")
-#' 
+#'
 #' # combining the plot with a title and a caption
 #' combine_plots(
 #'   p1,