+++ R/ggpiestats.R
@@ -71,7 +71,7 @@
 #' @importFrom paletteer scale_fill_paletteer_d
 #'
 #' @references
-#' \url{https://indrajeetpatil.github.io/ggstatsplot/articles/ggpiestats.html}
+#' \url{https://cran.r-project.org/web/packages/ggstatsplot/vignettes/ggpiestats.html}
 #'
 #' @examples
 #'