+++ R/grouped_gghistostats.R
@@ -32,10 +32,10 @@
 #' @seealso \code{\link{gghistostats}}
 #'
 #' @references
-#' \url{https://indrajeetpatil.github.io/ggstatsplot/articles/gghistostats.html}
+#' \url{https://cran.r-project.org/web/packages/ggstatsplot/vignettes/gghistostats.html}
 #'
 #' @examples
-#' 
+#'
 #' ggstatsplot::grouped_gghistostats(
 #'   data = iris,
 #'   x = Sepal.Length,