+++ R/grouped_ggscatterstats.R
@@ -33,13 +33,13 @@
 #' @seealso \code{\link{ggscatterstats}} \code{\link{ggcorrmat}} \code{\link{grouped_ggcorrmat}}
 #'
 #' @references
-#' \url{https://indrajeetpatil.github.io/ggstatsplot/articles/ggscatterstats.html}
+#' \url{https://cran.r-project.org/web/packages/ggstatsplot/vignettes/ggscatterstats.html}
 #'
 #' @examples
-#' 
+#'
 #' # to ensure reproducibility
 #' set.seed(123)
-#' 
+#'
 #' # basic functio call
 #' ggstatsplot::grouped_ggscatterstats(
 #'   data = datasets::iris,